Struct ruru::Array

pub struct Array { /* fields omitted */ }

Array

Methods

impl Array

fn new() -> Self

Creates a new instance of empty Array.

Examples

use ruru::{Array, VM};

Array::new();

Ruby:

[]

fn with_capacity(capacity: usize) -> Self

Creates a new instance of empty Array with reserved space for capacity elements.

Examples

use ruru::{Array, Fixnum, VM};

let mut array = Array::with_capacity(2);

assert_eq!(array.length(), 0);

array.push(Fixnum::new(1));
array.push(Fixnum::new(2));

assert_eq!(array.length(), 2);

Ruby:

[]

fn length(&self) -> usize

Retrieves the length of the array.

Examples

use ruru::{Array, Fixnum, VM};

let mut array = Array::new().push(Fixnum::new(1));

assert_eq!(array.length(), 1);

array.push(Fixnum::new(2));

assert_eq!(array.length(), 2);

Ruby:

array = [1]
array.length == 1

array << 2
array.length == 2

fn at(&self, index: i64) -> AnyObject

Retrieves an AnyObject from the element at index position.

Examples

use ruru::{Array, Fixnum, Object, VM};

let array = Array::new().push(Fixnum::new(1));

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));

Ruby:

array = [1]

array[0] == 1

fn join(&self, separator: RString) -> RString

Joins all elements of Array to Ruby String.

Examples

use ruru::{Array, RString, VM};

let array = Array::new()
    .push(RString::new("Hello"))
    .push(RString::new("World!"));

let joined_string = array.join(RString::new(", ")).to_string();

assert_eq!(joined_string, "Hello, World!".to_string());

Ruby:

array = ['Hello', 'World!']

joined_string = array.join(', ')

joined_string == 'Hello, World!'

fn push<T: Object>(&mut self, item: T) -> Self

Pushes an object to Array.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new();

array.push(Fixnum::new(1));

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));

Ruby:

array = []
array << 1

array[0] == 1

fn store<T: Object>(&mut self, index: i64, item: T) -> AnyObject

Stores an object at index position.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1));

array.store(0, Fixnum::new(2));

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [1]
array[0] = 2

array[0] == 2

fn pop(&mut self) -> AnyObject

Removes and returns the last element of the array.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1));

assert_eq!(array.pop().try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));

Ruby:

array = [1]

array.pop == 1

fn unshift<T: Object>(&mut self, item: T) -> Array

Inserts item at the beggining of the array.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1));

array.unshift(Fixnum::new(2));

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [1]
array.unshift(2)

array[0] == 2

fn shift(&mut self) -> AnyObject

Removes the first item of the array and returns it.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1)).push(Fixnum::new(2));

let item = array.shift();

assert_eq!(item.try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));
assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [1, 2]

item = array.shift

item == 1
array[0] == 2

fn dup(&self) -> Array

Creates a copy of the array.

Examples

use ruru::{Array, Fixnum, Object, VM};

let array = Array::new().push(Fixnum::new(1));
let copy = array.dup();

assert_eq!(array.at(0), copy.at(0));

Ruby:

array = [1]
copy = array.dup

array[0] == copy[0]

fn to_s(&self) -> RString

Creates a string representation of the array.

Examples

use ruru::{Array, Fixnum, VM};

let array = Array::new().push(Fixnum::new(1)).push(Fixnum::new(2));

assert_eq!(array.to_s().to_string(), "[1, 2]".to_string());

Ruby:

array = [1, 2]

array.to_s == "[1, 2]"

fn reverse(&self) -> Array

Returns a new array containing array's elements in reverse order.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1)).push(Fixnum::new(2));

let reversed_array = array.reverse();

assert_eq!(reversed_array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));
assert_eq!(reversed_array.at(1).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));

Ruby:

array = [1, 2]

reversed_array = array.reverse

reversed_array[0] == 2
reversed_array[1] == 1

fn reverse_bang(&mut self) -> Array

Reverses self in place.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1)).push(Fixnum::new(2));

array.reverse_bang();

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));
assert_eq!(array.at(1).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));

Ruby:

array = [1, 2]

array.reverse!

array[0] == 2
array[1] == 1

fn concat(&mut self, other: &Array) -> Array

Appends the elements of other array to self.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(1));
let other = Array::new().push(Fixnum::new(2));

array.concat(&other);

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));
assert_eq!(array.at(1).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [1]
other = [2]

array.concat(other)

array[0] == 1
array[1] == 2

fn sort(&self) -> Array

Returns a new array created by sorting self.

Examples

use ruru::{Array, Fixnum, Object, VM};

let array = Array::new().push(Fixnum::new(2)).push(Fixnum::new(1));

let sorted_array = array.sort();

assert_eq!(sorted_array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));
assert_eq!(sorted_array.at(1).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [2, 1]

sorted_array = array.sort

sorted_array[0] == 1
sorted_array[1] == 2

fn sort_bang(&mut self) -> Array

Sorts the array in place.

Examples

use ruru::{Array, Fixnum, Object, VM};

let mut array = Array::new().push(Fixnum::new(2)).push(Fixnum::new(1));

array.sort_bang();

assert_eq!(array.at(0).try_convert_to::<Fixnum>(), Ok(Fixnum::new(1)));
assert_eq!(array.at(1).try_convert_to::<Fixnum>(), Ok(Fixnum::new(2)));

Ruby:

array = [2, 1]

array.sort!

array[0] == 1
array[1] == 2