Struct ruru::RString

pub struct RString {
    // some fields omitted
}

String

Methods

impl RString

fn new(string: &str) -> Self

Creates a new instance of Ruby String containing given string.

Examples

use ruru::{RString, VM};

let string = RString::new("Hello, World!");

assert_eq!(string.to_string(), "Hello, World!".to_string());

Ruby:

str = 'Hello, World!'

str == 'Hello, World!'

fn to_string(&self) -> String

Retrieves underlying Rust String from Ruby String object.

Examples

use ruru::{RString, VM};

let string = RString::new("Hello, World!");

assert_eq!(string.to_string(), "Hello, World!".to_string());

Ruby:

str = 'Hello, World!'

str == 'Hello, World!'

fn to_string_unchecked(&self) -> String

Retrieves underlying Rust String from Ruby String object.

Unlike to_string() it does not perform any checks for internal null-bytes.

This function may be used to safely get binary data from Ruby.

Examples

use ruru::{RString, VM};

let string = RString::new("Hello,\0World!");

assert_eq!(string.to_string_unchecked(), "Hello,\0World!".to_string());

Ruby:

str = 'Hello,\0World!'

str == 'Hello,\0World!'

fn bytesize(&self) -> i64

Returns the length of the string in bytes

Examples

use ruru::{RString, VM};

let string = RString::new("Hello, World!");
let utf8_string = RString::new("⓯");

assert_eq!(string.bytesize(), 13);
assert_eq!(utf8_string.bytesize(), 3);

Ruby:

string = 'Hello, World!'
utf8_string = '⓯'

string.bytesize == 13
utf8_string.bytesize == 3