Struct ruru::Class

pub struct Class { /* fields omitted */ }

Class

Also see def, def_self, define and some more functions from Object trait.

#[macro_use] extern crate ruru;

use std::error::Error;

use ruru::{Class, Fixnum, Object, VM};

methods!(
   Fixnum,
   itself,

    fn pow(exp: Fixnum) -> Fixnum {
        // `exp` is not a valid `Fixnum`, raise an exception
        if let Err(ref error) = exp {
            VM::raise(error.to_exception(), error.description());
        }

        // We can safely unwrap here, because an exception was raised if `exp` is `Err`
        let exp = exp.unwrap().to_i64() as u32;

        Fixnum::new(itself.to_i64().pow(exp))
    }
);

fn main() {
    Class::from_existing("Fixnum").define(|itself| {
        itself.def("pow", pow);
    });
}

Ruby:

class Fixnum
  def pow(exp)
    raise TypeError unless exp.is_a?(Fixnum)

    self ** exp
  end
end

Methods

impl Class

fn new(name: &str, superclass: Option<&Self>) -> Self

Creates a new Class.

superclass can receive the following values:

None to inherit from Object class (standard Ruby behavior when superclass is not given explicitly);

Some(&Class) to inherit from the given class

Examples

use ruru::{Class, VM};

let basic_record_class = Class::new("BasicRecord", None);

assert_eq!(basic_record_class, Class::from_existing("BasicRecord"));
assert_eq!(basic_record_class.superclass(), Some(Class::from_existing("Object")));

let record_class = Class::new("Record", Some(&basic_record_class));

assert_eq!(record_class, Class::from_existing("Record"));
assert_eq!(record_class.superclass(), Some(Class::from_existing("BasicRecord")));

Ruby:

class BasicRecord
end

class Record < BasicRecord
end

BasicRecord.superclass == Object

Record.superclass == BasicRecord

fn from_existing(name: &str) -> Self

Retrieves an existing Class object.

Examples

use ruru::{Class, VM};

let class = Class::new("Record", None);

assert_eq!(class, Class::from_existing("Record"));

Ruby:

class Record
end

# get class

Record

# or

Object.const_get('Record')

fn new_instance(&self, arguments: Vec<AnyObject>) -> AnyObject

Creates a new instance of Class

Arguments must be passed as a vector of AnyObject (see example).

Examples

use ruru::{Class, Fixnum, Object};

// Without arguments
Class::from_existing("Hello").new_instance(vec![]);

// With arguments passing arguments to constructor
let arguments = vec![
    Fixnum::new(1).to_any_object(),
    Fixnum::new(2).to_any_object()
];

Class::from_existing("Worker").new_instance(arguments);

Ruby:

Hello.new

Worker.new(1, 2)

fn superclass(&self) -> Option<Class>

Returns a superclass of the current class

Examples

use ruru::{Class, Object, VM};

assert_eq!(
    Class::from_existing("Array").superclass(),
    Some(Class::from_existing("Object"))
);

assert_eq!(Class::from_existing("BasicObject").superclass(), None);

fn ancestors(&self) -> Vec<Class>

Returns a Vector of ancestors of current class

Examples

Getting all the ancestors

use ruru::{Class, VM};

let true_class_ancestors = Class::from_existing("TrueClass").ancestors();

let expected_ancestors = vec![
    Class::from_existing("TrueClass"),
    Class::from_existing("Object"),
    Class::from_existing("Kernel"),
    Class::from_existing("BasicObject")
];

assert_eq!(true_class_ancestors, expected_ancestors);

Searching for an ancestor

use ruru::{Class, VM};

let basic_record_class = Class::new("BasicRecord", None);
let record_class = Class::new("Record", Some(&basic_record_class));

let ancestors = record_class.ancestors();

assert!(ancestors.iter().any(|class| *class == basic_record_class));

fn get_nested_class(&self, name: &str) -> Self

Retrieves a Class nested to current Class.

Examples

use ruru::{Class, Object, VM};

Class::new("Outer", None).define(|itself| {
    itself.define_nested_class("Inner", None);
});

Class::from_existing("Outer").get_nested_class("Inner");

Ruby:

class Outer
  class Inner
  end
end

Outer::Inner

# or

Outer.const_get('Inner')

fn define_nested_class(&mut self,
name: &str,
superclass: Option<&Class>)
-> Self

Creates a new Class nested into current class.

superclass can receive the following values:

None to inherit from Object class (standard Ruby behavior when superclass is not given explicitly);

Some(&class) to inherit from the given class

Examples

use ruru::{Class, Object, VM};

Class::new("Outer", None).define(|itself| {
    itself.define_nested_class("Inner", None);
});

Class::from_existing("Outer").get_nested_class("Inner");

Ruby:

class Outer
  class Inner
  end
end

Outer::Inner

# or

Outer.const_get('Inner')

fn const_get(&self, name: &str) -> AnyObject

Retrieves a constant from class.

Examples

use ruru::{Class, Object, RString, VM};

Class::new("Greeter", None).define(|itself| {
    itself.const_set("GREETING", &RString::new("Hello, World!"));
});

let greeting = Class::from_existing("Greeter")
    .const_get("GREETING")
    .try_convert_to::<RString>().unwrap()
    .to_string();

assert_eq!(greeting, "Hello, World!");

Ruby:

class Greeter
  GREETING = 'Hello, World!'
end

# or

Greeter = Class.new
Greeter.const_set('GREETING', 'Hello, World!')

# ...

Greeter::GREETING == 'Hello, World!'

# or

Greeter.const_get('GREETING') == 'Hello, World'

fn const_set<T: Object>(&mut self, name: &str, value: &T)

Defines a constant for class.

Examples

use ruru::{Class, Object, RString, VM};

Class::new("Greeter", None).define(|itself| {
    itself.const_set("GREETING", &RString::new("Hello, World!"));
});

let greeting = Class::from_existing("Greeter")
    .const_get("GREETING")
    .try_convert_to::<RString>().unwrap()
    .to_string();

assert_eq!(greeting, "Hello, World!");

Ruby:

class Greeter
  GREETING = 'Hello, World!'
end

# or

Greeter = Class.new
Greeter.const_set('GREETING', 'Hello, World!')

# ...

Greeter::GREETING == 'Hello, World!'

# or

Greeter.const_get('GREETING') == 'Hello, World'

fn attr_reader(&mut self, name: &str)

Defines an attr_reader for class

Examples

use ruru::{Class, Object, VM};

Class::new("Test", None).define(|itself| {
    itself.attr_reader("reader");
});

Ruby:

class Test
  attr_reader :reader
end

fn attr_writer(&mut self, name: &str)

Defines an attr_writer for class

Examples

use ruru::{Class, Object, VM};

Class::new("Test", None).define(|itself| {
    itself.attr_writer("writer");
});

Ruby:

class Test
  attr_writer :writer
end

fn attr_accessor(&mut self, name: &str)

Defines an attr_accessor for class

Examples

use ruru::{Class, Object, VM};

Class::new("Test", None).define(|itself| {
    itself.attr_accessor("accessor");
});

Ruby:

class Test
  attr_accessor :accessor
end

fn wrap_data<T, O: Object>(&self, data: T, wrapper: &DataTypeWrapper<T>) -> O

Wraps Rust structure into a new Ruby object of the current class.

See the documentation for wrappable_struct! macro for more information.

Examples

Wrap Server structs to RubyServer objects

#[macro_use] extern crate ruru;
#[macro_use] extern crate lazy_static;

use ruru::{AnyObject, Class, Fixnum, Object, RString, VM};

// The structure which we want to wrap
pub struct Server {
    host: String,
    port: u16,
}

impl Server {
    fn new(host: String, port: u16) -> Self {
        Server {
            host: host,
            port: port,
        }
    }

    fn host(&self) -> &str {
        &self.host
    }

    fn port(&self) -> u16 {
        self.port
    }
}

wrappable_struct!(Server, ServerWrapper, SERVER_WRAPPER);

class!(RubyServer);

methods!(
    RubyServer,
    itself,

    fn ruby_server_new(host: RString, port: Fixnum) -> AnyObject {
        let server = Server::new(host.unwrap().to_string(),
                                 port.unwrap().to_i64() as u16);

        Class::from_existing("RubyServer").wrap_data(server, &*SERVER_WRAPPER)
    }

    fn ruby_server_host() -> RString {
        let host = itself.get_data(&*SERVER_WRAPPER).host();

        RString::new(host)
    }

    fn ruby_server_port() -> Fixnum {
        let port = itself.get_data(&*SERVER_WRAPPER).port();

        Fixnum::new(port as i64)
    }
);

fn main() {
    let data_class = Class::from_existing("Data");

    Class::new("RubyServer", Some(&data_class)).define(|itself| {
        itself.def_self("new", ruby_server_new);

        itself.def("host", ruby_server_host);
        itself.def("port", ruby_server_port);
    });
}

To use the RubyServer class in Ruby:

server = RubyServer.new("127.0.0.1", 3000)

server.host == "127.0.0.1"
server.port == 3000